Migrated ServerModMetadata type to ServerMod.cs
