--- project/SamSWAT.FireSupport.Server/ServerModMetadata.cs
@@ -1,20 +1,0 @@
-﻿using SPTarkov.Server.Core.Models.Spt.Mod;
-using Range = SemanticVersioning.Range;
-using Version = SemanticVersioning.Version;
-
-namespace SamSWAT.FireSupport.ArysReloaded;
-
-public record ServerModMetadata : AbstractModMetadata
-{
-	public override string ModGuid { get; init; } = "com.samswat.firesupport.arysreloaded";
-	public override string Name { get; init; } = "SamSWAT's FireSupport: Arys Reloaded";
-	public override string Author { get; init; } = "SamSWAT, Arys";
-	public override List<string>? Contributors { get; init; }
-	public override Version Version { get; init; } = new(ModMetadata.VERSION);
-	public override Range SptVersion { get; init; } = new($"~{ModMetadata.TARGET_SPT_VERSION}");
-	public override List<string>? Incompatibilities { get; init; }
-	public override Dictionary<string, Range>? ModDependencies { get; init; }
-	public override string? Url { get; init; }
-	public override bool? IsBundleMod { get; init; }
-	public override string License { get; init; } = "Creative Commons BY-NC 3.0";
-}
